refactor(proxy): streamline proxy authentication handling and improve code readability
